Реализует открытие контекстного меню сортировки.
