+++ src/components/sort-type.tsx
@@ -1,22 +1,11 @@
-import { useAppDispatch } from '../hooks';
-import { changeSortType, sortCityOffers } from '../store/actions';
-
 type SortTypeProps = {
   sortTypeName: string;
   isSelected: boolean;
+  onTypeSelected: (sortTypeName: string, isSelected: boolean) => void;
 }
 
-export default function SortType ({sortTypeName, isSelected}: SortTypeProps): JSX.Element {
-  const dispatch = useAppDispatch();
-
-  function handleClick () {
-    if (!isSelected) {
-      dispatch(changeSortType({sortTypeName}));
-      dispatch(sortCityOffers());
-    }
-  }
-
+export default function SortType ({sortTypeName, isSelected, onTypeSelected}: SortTypeProps): JSX.Element {
   return (
-    <li className={`places__option ${isSelected ? 'places__option--active' : null} tabIndex={0}`} onClick={handleClick}>{sortTypeName}</li>
+    <li className={`places__option ${isSelected ? 'places__option--active' : null} tabIndex={0}`} onClick={() => onTypeSelected(sortTypeName, isSelected)}>{sortTypeName}</li>
   );
 }
